Cleaned up default value and added round bars
Made it so the text fields in the GUI are default to the constants set
in the main class.

Added two round bar renderers. One is filled and one is outlined. Both
are drawn exact and anti-aliased (but the standard block renders aren't
so they will take less time to render than the rounds)

Dakkra